server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bson==0.5.10
pymongo==3.11.0
python-dotenv==0.21.0
